Code cleanup, example notebook, requirements
