+++ ceruleanserver/associate/requirements.txt
@@ -1,0 +1,11 @@
+centerline==1.0.1
+earthengine-api==0.1.342
+geopandas==0.12.2
+ipyleaflet==0.17.2
+ipywidgets==8.0.4
+matplotlib==3.7.0
+movingpandas==0.15rc1
+numpy==1.21.5
+pyproj==3.4.1
+scipy==1.8.0
+shapely==2.0.1
