+++ ceruleanserver/associate/sample_data/S1B_IW_GRDH_1SDV_20210626T051247_20210626T051312_027523_034911_53D5_oldmodel.csv
@@ -1,0 +1,2 @@
+ssvid,coinc_score
+215839000,0.10359762717920976
